Add integration tests for mgit CLI commands and update PyInstaller spec
- Created a PyInstaller spec file for packaging mgit.
- Added pip-audit and safety reports for dependency vulnerability checks.
- Implemented integration tests for authentication commands, configuration commands, repository commands, and status command.
- Tests include scenarios for successful login, invalid credentials, environment variable usage, configuration display, and repository cloning/pulling.
- Enhanced test coverage for handling existing repositories and error scenarios during git operations.
- Updated test cases to ensure proper output and exit codes for various command scenarios.
