+++ .clinerules/03-code-modularity.md
@@ -1,0 +1,7 @@
+## Brief overview
+This rule defines a guideline for maintaining code modularity by monitoring file sizes.
+
+## Code Modularity
+- When a source code file exceeds 300 lines of code, it should be flagged for review.
+- The review should assess opportunities to break the file down into smaller, more focused modules.
+- This helps to improve readability, maintainability, and testability.
