+++ .gitignore
@@ -13,7 +13,6 @@ dist/
 
 # PyInstaller
 *.manifest
-*.spec
 
 # Testing
 htmlcov/
