+++ mgit/commands/monitoring.py
@@ -26,7 +26,7 @@ from ..monitoring.server import start_monitoring_server, start_simple_monitoring
 
 
 def start_monitoring_server_command(
-    host: str = "0.0.0.0",
+    host: str = "127.0.0.1",
     port: int = 8080,
     simple: bool = False,
     log_level: str = "INFO",
@@ -316,7 +316,7 @@ if TYPER_AVAILABLE:
 
     @monitoring_app.command("server")
     def server_cmd(
-        host: str = typer.Option("0.0.0.0", "--host", "-h", help="Server host"),
+        host: str = typer.Option("127.0.0.1", "--host", "-h", help="Server host"),
         port: int = typer.Option(8080, "--port", "-p", help="Server port"),
         simple: bool = typer.Option(
             False, "--simple", help="Use simple server (no aiohttp)"
