+++ pip-audit-report.json
@@ -1,0 +1,1 @@
+{"dependencies": [{"name": "aiohappyeyeballs", "version": "2.6.1", "vulns": []}, {"name": "aiohttp", "version": "3.12.11", "vulns": []}, {"name": "aiosignal", "version": "1.3.2", "vulns": []}, {"name": "altgraph", "version": "0.17.4", "vulns": []}, {"name": "annotated-types", "version": "0.7.0", "vulns": []}, {"name": "attrs", "version": "25.3.0", "vulns": []}, {"name": "authlib", "version": "1.6.0", "vulns": []}, {"name": "azure-core", "version": "1.34.0", "vulns": []}, {"name": "azure-devops", "version": "7.1.0b4", "vulns": []}, {"name": "bandit", "version": "1.8.3", "vulns": []}, {"name": "black", "version": "24.10.0", "vulns": []}, {"name": "boolean-py", "version": "5.0", "vulns": []}, {"name": "cachecontrol", "version": "0.14.3", "vulns": []}, {"name": "certifi", "version": "2025.4.26", "vulns": []}, {"name": "cffi", "version": "1.17.1", "vulns": []}, {"name": "charset-normalizer", "version": "3.4.2", "vulns": []}, {"name": "click", "version": "8.1.8", "vulns": []}, {"name": "coverage", "version": "7.8.2", "vulns": []}, {"name": "cryptography", "version": "45.0.3", "vulns": []}, {"name": "cyclonedx-python-lib", "version": "9.1.0", "vulns": []}, {"name": "defusedxml", "version": "0.7.1", "vulns": []}, {"name": "dparse", "version": "0.6.4", "vulns": []}, {"name": "filelock", "version": "3.12.4", "vulns": []}, {"name": "frozenlist", "version": "1.6.2", "vulns": []}, {"name": "idna", "version": "3.10", "vulns": []}, {"name": "iniconfig", "version": "2.1.0", "vulns": []}, {"name": "isodate", "version": "0.7.2", "vulns": []}, {"name": "jinja2", "version": "3.1.6", "vulns": []}, {"name": "license-expression", "version": "30.4.1", "vulns": []}, {"name": "markdown-it-py", "version": "3.0.0", "vulns": []}, {"name": "markupsafe", "version": "3.0.2", "vulns": []}, {"name": "marshmallow", "version": "4.0.0", "vulns": []}, {"name": "mdurl", "version": "0.1.2", "vulns": []}, {"name": "memory-profiler", "version": "0.61.0", "vulns": []}, {"name": "mgit", "skip_reason": "Dependency not found on PyPI and could not be audited: mgit (0.2.13)"}, {"name": "msgpack", "version": "1.1.0", "vulns": []}, {"name": "msrest", "version": "0.7.1", "vulns": []}, {"name": "multidict", "version": "6.4.4", "vulns": []}, {"name": "mypy", "version": "1.16.0", "vulns": []}, {"name": "mypy-extensions", "version": "1.1.0", "vulns": []}, {"name": "oauthlib", "version": "3.2.2", "vulns": []}, {"name": "packageurl-python", "version": "0.17.1", "vulns": []}, {"name": "packaging", "version": "25.0", "vulns": []}, {"name": "pastel", "version": "0.2.1", "vulns": []}, {"name": "pathspec", "version": "0.12.1", "vulns": []}, {"name": "pbr", "version": "6.1.1", "vulns": []}, {"name": "pip", "version": "25.1.1", "vulns": []}, {"name": "pip-api", "version": "0.0.34", "vulns": []}, {"name": "pip-audit", "version": "2.9.0", "vulns": []}, {"name": "pip-requirements-parser", "version": "32.0.1", "vulns": []}, {"name": "platformdirs", "version": "4.3.8", "vulns": []}, {"name": "pluggy", "version": "1.6.0", "vulns": []}, {"name": "poethepoet", "version": "0.24.4", "vulns": []}, {"name": "propcache", "version": "0.3.1", "vulns": []}, {"name": "psutil", "version": "6.0.0", "vulns": []}, {"name": "py-cpuinfo", "version": "9.0.0", "vulns": []}, {"name": "py-serializable", "version": "2.0.0", "vulns": []}, {"name": "pycparser", "version": "2.22", "vulns": []}, {"name": "pydantic", "version": "2.9.2", "vulns": []}, {"name": "pydantic-core", "version": "2.23.4", "vulns": []}, {"name": "pygments", "version": "2.19.1", "vulns": []}, {"name": "pyinstaller", "version": "6.14.1", "vulns": []}, {"name": "pyinstaller-hooks-contrib", "version": "2025.5", "vulns": []}, {"name": "pyparsing", "version": "3.2.3", "vulns": []}, {"name": "pytest", "version": "7.4.4", "vulns": []}, {"name": "pytest-asyncio", "version": "0.21.2", "vulns": []}, {"name": "pytest-benchmark", "version": "4.0.0", "vulns": []}, {"name": "pytest-cov", "version": "4.1.0", "vulns": []}, {"name": "python-dotenv", "version": "1.1.0", "vulns": []}, {"name": "pyyaml", "version": "6.0.2", "vulns": []}, {"name": "requests", "version": "2.32.3", "vulns": []}, {"name": "requests-oauthlib", "version": "2.0.0", "vulns": []}, {"name": "rich", "version": "13.9.4", "vulns": []}, {"name": "ruamel-yaml", "version": "0.18.13", "vulns": []}, {"name": "ruamel-yaml-clib", "version": "0.2.12", "vulns": []}, {"name": "ruff", "version": "0.1.15", "vulns": []}, {"name": "safety", "version": "3.2.13", "vulns": []}, {"name": "safety-schemas", "version": "0.0.10", "vulns": []}, {"name": "setuptools", "version": "80.9.0", "vulns": []}, {"name": "six", "version": "1.17.0", "vulns": []}, {"name": "sortedcontainers", "version": "2.4.0", "vulns": []}, {"name": "stevedore", "version": "5.4.1", "vulns": []}, {"name": "toml", "version": "0.10.2", "vulns": []}, {"name": "tomli", "version": "2.2.1", "vulns": []}, {"name": "typer", "version": "0.9.4", "vulns": []}, {"name": "typing-extensions", "version": "4.14.0", "vulns": []}, {"name": "urllib3", "version": "2.4.0", "vulns": []}, {"name": "yarl", "version": "1.20.0", "vulns": []}], "fixes": []}
