+++ pyproject.toml
@@ -21,7 +21,7 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.9,<3.13"
+python = ">=3.9.2,<3.13"
 aiohttp = "^3.9.0"
 azure-devops = "^7.1.0b1"
 python-dotenv = "^1.0.0"
@@ -29,6 +29,7 @@ rich = "^13.0.0"
 typer = "^0.9.0"
 pyyaml = "^6.0"
 "ruamel.yaml" = "^0.18.0"
+cryptography = ">=44.0.1"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.0.0"
